requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-playwright==0.3.0
playwright==1.40.0

# Web3 and Circle SDK
web3==7.14.0
circle-configurations==8.1.1
circle-developer-controlled-wallets==8.1.1
circle-web3-sdk-util==8.1.1

# OpenAI
openai==1.3.5

# Environment management
python-dotenv==1.0.0

# Database
aiosqlite==0.19.0

# Utilities
python-multipart==0.0.6
httpx==0.25.1
requests==2.31.0
elevenlabs==0.2.0
# <parameter name="filePath">/Users/enkay/src/dev/nsurecat-genesis001/requirements.txt